Most of the remaining pre-defined symbol ids for expressions.
It seems like symbol ids and instruction ids don't overlap at all. If that's truly the case, the only symbols left are ids 0x50 and 0x51.

The ones added here are all of the bitwise operations, bit shifting, and the modulo operator.

All of these operators have been tested in a custom-made script so I'm fairly certain they're right